Add library image for registry
This is based on swarm-library-image.

Unfortunately, the registry binary needs to be checked in - otherwise
"docker build" wouldn't work without preparation, as is required of
official images. The binary checked in here is from distribution's
current master branch (4cc4d440f6afc7b27c2220993d01a4e2889d87c0). It
will need to be updated to the release version after each release.

Rather than creating a container from scratch, this library image is
based on the ubuntu container, because registry is a dynamically linked
binary.

Because we're using the ubuntu container as a base, manual installation
of ca-certificates.crt isn't necessary. If we end up switching to a
scratch container, this will need to be re-added.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>